调整存储数据结构，完成初始化基础数据加载

+++ Assets/Scripts/MainLogic/Level/Unit/UnitData.cs
@@ -53,7 +53,7 @@ public class UnitData
     /// <summary>
     /// 命中率
     /// </summary>
-    public int HitChance;
+    public float HitChance;
     /// <summary>
     /// 闪避率
     /// </summary>
